Django/myproject/requirements.txt
certifi==2016.2.28
colorama==0.3.9
decorator==4.1.2
Django==1.11.6
imageio==2.2.0
ipython==6.1.0
ipython-genutils==0.2.0
jedi==0.10.2
numpy==1.13.1
olefile==0.44
path.py==10.3.1
pickleshare==0.7.4
Pillow==4.2.1
prompt-toolkit==1.0.15
Pygments==2.2.0
pytz==2017.2
simplegeneric==0.8.1
six==1.10.0
traitlets==4.3.2
wcwidth==0.1.7
wincertstore==0.2
